electricity billing system project/solar_bill_ai/requirements.txt
streamlit
pytesseract
pillow
openpyxl
pandas
